Implement redirection functionality
A basic implementation of the functionality that allows a shortened URL to redirect a user to the resource at the original address.  In addition, some of the unnecessary methods that were previously used in development testing have been removed, and separate models that can be used by the API and UI have been added for transferring data.