Remove the API URL from the UI project code
The API URL was hard-coded in the UI project files.  Although this was acceptable for development testing, the API URL could change depending on the environment, and so storing the URL as an application setting is much more robust solution.  A production URL has not yet been determined, and so is left blank for now.