Add input validation
Previously, the application did not perform any checks on the input that was provided by the user; neither the values passed to the user interface nor the values passed to the API were validated.
This commit applies validation to both the front- and back-end logic to ensure that the input loosely adheres to the requirements of a URL.  Validation messages have been added to the user interface as well.
In addition, the algorithm that creates shortened URL identifiers has been updated; the generated identifiers are now shorter because only part of the year is now used.